Changed tasks to the improved UNet and began developing basic structure based off provided research paper.
